Add files via upload
I made some modifications to the little demo I had last class, I wasn't all happy with the movement of the ship in terms of the values being stacked for the ship movement. The code is much cleaner and simpler now, I also added a substitute for the background and ship as shapes so it can run without downloading any additional files. feel free to play with the code or comment any ideas. I have been getting into generating asteroids to test collision and breakdown of said asteroids into smaller elements. For now, nothing too interesting, just playing around to get a better feel of things so we can work on those things later on.