Add files via upload
Making a new branch just to keep files updated. After lots of headaches and reading/playing with things, I managed to get a version of the same experiment to work using JOGL. Now the FrameAnimator can take care of the instructions per second so we won't need the thread sleep anymore. Rotation for a simple figure has been implemented, and I have gone through some very very basic 3d modeling with polygons (not for this code, just me playing with cubes and stuff). I have some ideas on how to get rid of the JFrame and just use a GL window so that it's all in open GL and not a hybrid between both. (implementing the key listeners was a bit strange with the hybrid). I will mess with this idea a little more after today's assignments and will upload a cleaner version of this that we can actually start messing around with! Feel free to test this out, you will need to have JOGL with eclipse or other compiler for this to work.

PS.
Long live the windows 95 style~ Good testing colors.