Add files via upload
removed the large sound file for BGM, just replace with any previous version of bgm. Also, ALL MODELS ARE NEW. so make sure you replace all files with the ones in this package :) oh, and remember I use the "filepath\\item.file" so make sure to change that if needed.